refactor: 로그인, 상단바 로고 변경
